requirements.txt
streamlit
PyPDF2
python-docx
requests
python-dotenv